+++ src/core/app/builder.rs
@@ -5,6 +5,7 @@ use std::fmt::{Debug};
 use std::sync::{Arc, Mutex};
 use to_mut_proc_macro::ToMut;
 use to_mut::ToMut;
+use tracing_subscriber;
 use clap::{Arg, ArgAction, Command as ClapCommand};
 use dotenvy::dotenv;
 use crate::connectors::mongodb::connector::MongoDBConnector;
@@ -254,6 +255,13 @@ impl AppBuilder {
         let source = parser.get_source(connector_ref.0);
         let connector_declaration = source.get_connector(connector_ref.1);
         let url = connector_declaration.url.as_ref().unwrap();
+        if connector_declaration.debug {
+            env::set_var("FMT_SQL", "true");
+            // set tracing
+            // install global collector configured based on RUST_LOG env var.
+            env::set_var("RUST_LOG", "none,quaint=debug");
+            tracing_subscriber::fmt::init();
+        }
         let connector: Arc<dyn Connector> = match connector_declaration.provider.unwrap() {
             DatabaseName::MySQL => {
                 #[cfg(feature = "data-source-mysql")]